requirements.txt
asyncpg
fastapi
python-jose[cryptography]
pydantic
sqlalchemy
python-dotenv
uvicorn[standard]
